requirements.txt
pip
setuptools==56.1.0
Django==4.0
Pillow==8.4.0
django-allauth
django-crispy-forms==1.13.0
django-cors-headers==3.10.1
djangorestframework-jwt